Restore registration

+++ app/controllers/registrations_controller.rb
@@ -1,4 +1,5 @@
 class RegistrationsController < Devise::RegistrationsController
+  # before_filter :configure_permitted_parameters, :only => [:create]
 
   def create
     # binding.b
@@ -7,17 +8,19 @@ class RegistrationsController < Devise::RegistrationsController
       resource.save
       yield resource if block_given?
       if resource.persisted?
-        @payment = Payment.new({ email: params["user"]["email"],
-                                 token: params[:payment]["token"], user_id: resource.id })
-        flash[:error] = "Please check registration errors" unless @payment.valid?
-        begin
-          @payment.process_payment
-          @payment.save
-        rescue Exception => e
-          flash[:error] = e.message
-          resource.destroy
-          puts 'Payment failed'
-          render :new and return
+        if params[:payment]
+          @payment = Payment.new({ email: params["user"]["email"],
+                                   token: params[:payment]["token"], user_id: resource.id })
+          flash[:error] = "Please check registration errors" unless @payment.valid?
+          begin
+            @payment.process_payment
+            @payment.save
+          rescue Exception => e
+            flash[:error] = e.message
+            resource.destroy
+            puts 'Payment failed'
+            render :new and return
+          end
         end
         if resource.active_for_authentication?
           set_flash_message :notice, :signed_up if is_flashing_format?
@@ -39,6 +42,6 @@ class RegistrationsController < Devise::RegistrationsController
   protected
 
   def configure_permitted_parameters
-    devise_parameter_sanitizer.for(:sign_up).push(:payment)
+    devise_parameter_sanitizer.permit(:sign_up, keys: [:email, :password, payment: [:card_number]])
   end
 end
